w3c-reports/12092014/gtk/runned-tests/html/syntax/parsing/html5lib_tests7-expected.txt

CONSOLE MESSAGE: Blocked a frame with origin "http://localhost:8800" from accessing a frame with origin "null".  The frame requesting access has a protocol of "http", the frame being accessed has a protocol of "data". Protocols must match.

CONSOLE MESSAGE: Blocked a frame with origin "http://localhost:8800" from accessing a frame with origin "null".  The frame requesting access has a protocol of "http", the frame being accessed has a protocol of "data". Protocols must match.

CONSOLE MESSAGE: Blocked a frame with origin "http://localhost:8800" from accessing a frame with origin "null".  The frame requesting access has a protocol of "http", the frame being accessed has a protocol of "data". Protocols must match.

CONSOLE MESSAGE: Blocked a frame with origin "http://localhost:8800" from accessing a frame with origin "null".  The frame requesting access has a protocol of "http", the frame being accessed has a protocol of "data". Protocols must match.

CONSOLE MESSAGE: Blocked a frame with origin "http://localhost:8800" from accessing a frame with origin "null".  The frame requesting access has a protocol of "http", the frame being accessed has a protocol of "data". Protocols must match.

CONSOLE MESSAGE: Blocked a frame with origin "http://localhost:8800" from accessing a frame with origin "null".  The frame requesting access has a protocol of "http", the frame being accessed has a protocol of "data". Protocols must match.

CONSOLE MESSAGE: Blocked a frame with origin "http://localhost:8800" from accessing a frame with origin "null".  The frame requesting access has a protocol of "http", the frame being accessed has a protocol of "data". Protocols must match.

CONSOLE MESSAGE: Blocked a frame with origin "http://localhost:8800" from accessing a frame with origin "null".  The frame requesting access has a protocol of "http", the frame being accessed has a protocol of "data". Protocols must match.

CONSOLE MESSAGE: line 1: SyntaxError: Unexpected token '<'
CONSOLE MESSAGE: Blocked a frame with origin "http://localhost:8800" from accessing a frame with origin "null".  The frame requesting access has a protocol of "http", the frame being accessed has a protocol of "data". Protocols must match.

CONSOLE MESSAGE: Blocked a frame with origin "http://localhost:8800" from accessing a frame with origin "null".  The frame requesting access has a protocol of "http", the frame being accessed has a protocol of "data". Protocols must match.

CONSOLE MESSAGE: Blocked a frame with origin "http://localhost:8800" from accessing a frame with origin "null".  The frame requesting access has a protocol of "http", the frame being accessed has a protocol of "data". Protocols must match.

CONSOLE MESSAGE: Blocked a frame with origin "http://localhost:8800" from accessing a frame with origin "null".  The frame requesting access has a protocol of "http", the frame being accessed has a protocol of "data". Protocols must match.

CONSOLE MESSAGE: Blocked a frame with origin "http://localhost:8800" from accessing a frame with origin "null".  The frame requesting access has a protocol of "http", the frame being accessed has a protocol of "data". Protocols must match.

CONSOLE MESSAGE: Blocked a frame with origin "http://localhost:8800" from accessing a frame with origin "null".  The frame requesting access has a protocol of "http", the frame being accessed has a protocol of "data". Protocols must match.

CONSOLE MESSAGE: Blocked a frame with origin "http://localhost:8800" from accessing a frame with origin "null".  The frame requesting access has a protocol of "http", the frame being accessed has a protocol of "data". Protocols must match.

CONSOLE MESSAGE: Blocked a frame with origin "http://localhost:8800" from accessing a frame with origin "null".  The frame requesting access has a protocol of "http", the frame being accessed has a protocol of "data". Protocols must match.

CONSOLE MESSAGE: Blocked a frame with origin "http://localhost:8800" from accessing a frame with origin "null".  The frame requesting access has a protocol of "http", the frame being accessed has a protocol of "data". Protocols must match.

CONSOLE MESSAGE: Blocked a frame with origin "http://localhost:8800" from accessing a frame with origin "null".  The frame requesting access has a protocol of "http", the frame being accessed has a protocol of "data". Protocols must match.

CONSOLE MESSAGE: Blocked a frame with origin "http://localhost:8800" from accessing a frame with origin "null".  The frame requesting access has a protocol of "http", the frame being accessed has a protocol of "data". Protocols must match.

CONSOLE MESSAGE: Blocked a frame with origin "http://localhost:8800" from accessing a frame with origin "null".  The frame requesting access has a protocol of "http", the frame being accessed has a protocol of "data". Protocols must match.

CONSOLE MESSAGE: Blocked a frame with origin "http://localhost:8800" from accessing a frame with origin "null".  The frame requesting access has a protocol of "http", the frame being accessed has a protocol of "data". Protocols must match.

CONSOLE MESSAGE: Blocked a frame with origin "http://localhost:8800" from accessing a frame with origin "null".  The frame requesting access has a protocol of "http", the frame being accessed has a protocol of "data". Protocols must match.

CONSOLE MESSAGE: Blocked a frame with origin "http://localhost:8800" from accessing a frame with origin "null".  The frame requesting access has a protocol of "http", the frame being accessed has a protocol of "data". Protocols must match.

CONSOLE MESSAGE: Blocked a frame with origin "http://localhost:8800" from accessing a frame with origin "null".  The frame requesting access has a protocol of "http", the frame being accessed has a protocol of "data". Protocols must match.

CONSOLE MESSAGE: Blocked a frame with origin "http://localhost:8800" from accessing a frame with origin "null".  The frame requesting access has a protocol of "http", the frame being accessed has a protocol of "data". Protocols must match.

CONSOLE MESSAGE: Blocked a frame with origin "http://localhost:8800" from accessing a frame with origin "null".  The frame requesting access has a protocol of "http", the frame being accessed has a protocol of "data". Protocols must match.

CONSOLE MESSAGE: Blocked a frame with origin "http://localhost:8800" from accessing a frame with origin "null".  The frame requesting access has a protocol of "http", the frame being accessed has a protocol of "data". Protocols must match.

CONSOLE MESSAGE: Blocked a frame with origin "http://localhost:8800" from accessing a frame with origin "null".  The frame requesting access has a protocol of "http", the frame being accessed has a protocol of "data". Protocols must match.

CONSOLE MESSAGE: Blocked a frame with origin "http://localhost:8800" from accessing a frame with origin "null".  The frame requesting access has a protocol of "http", the frame being accessed has a protocol of "data". Protocols must match.

html5lib Parser Test


FAIL html5lib_tests7.html 7cb496e242a4dc9aed321252b5ca6ebf4f02ebcd null is not an object (evaluating 'element.normalize')(stack: test_serializer@http://localhost:8800/html/syntax/parsing/test.js:31:10
http://localhost:8800/html/syntax/parsing/test.js:162:52
step@http://localhost:8800/resources/testharness.js:798:30
onload@http://localhost:8800/html/syntax/parsing/test.js:160:13)
FAIL html5lib_tests7.html c0cffec1e999db2aefb2f6beb679fd9620566dbd null is not an object (evaluating 'element.normalize')(stack: test_serializer@http://localhost:8800/html/syntax/parsing/test.js:31:10
http://localhost:8800/html/syntax/parsing/test.js:162:52
step@http://localhost:8800/resources/testharness.js:798:30
onload@http://localhost:8800/html/syntax/parsing/test.js:160:13)
FAIL html5lib_tests7.html 7c644a6da21bfd551385b0a5044b82cf7be0a22f null is not an object (evaluating 'element.normalize')(stack: test_serializer@http://localhost:8800/html/syntax/parsing/test.js:31:10
http://localhost:8800/html/syntax/parsing/test.js:162:52
step@http://localhost:8800/resources/testharness.js:798:30
onload@http://localhost:8800/html/syntax/parsing/test.js:160:13)
FAIL html5lib_tests7.html 52fde917ba333b89afeff0e31104421455f4bf1b null is not an object (evaluating 'element.normalize')(stack: test_serializer@http://localhost:8800/html/syntax/parsing/test.js:31:10
http://localhost:8800/html/syntax/parsing/test.js:162:52
step@http://localhost:8800/resources/testharness.js:798:30
onload@http://localhost:8800/html/syntax/parsing/test.js:160:13)
FAIL html5lib_tests7.html a8f53ca779c0e5fc484771c4ec2aa6fb6d609779 null is not an object (evaluating 'element.normalize')(stack: test_serializer@http://localhost:8800/html/syntax/parsing/test.js:31:10
http://localhost:8800/html/syntax/parsing/test.js:162:52
step@http://localhost:8800/resources/testharness.js:798:30
onload@http://localhost:8800/html/syntax/parsing/test.js:160:13)
FAIL html5lib_tests7.html e4ce65a5fb6a3726b341ec94da583dee7c2c8232 null is not an object (evaluating 'element.normalize')(stack: test_serializer@http://localhost:8800/html/syntax/parsing/test.js:31:10
http://localhost:8800/html/syntax/parsing/test.js:162:52
step@http://localhost:8800/resources/testharness.js:798:30
onload@http://localhost:8800/html/syntax/parsing/test.js:160:13)
FAIL html5lib_tests7.html 8779e761986b4c724bfe73fee95b7972145fb4d3 null is not an object (evaluating 'element.normalize')(stack: test_serializer@http://localhost:8800/html/syntax/parsing/test.js:31:10
http://localhost:8800/html/syntax/parsing/test.js:162:52
step@http://localhost:8800/resources/testharness.js:798:30
onload@http://localhost:8800/html/syntax/parsing/test.js:160:13)
FAIL html5lib_tests7.html 620e44a8a55e82cec0d51e9d93025d8a5c4456fc null is not an object (evaluating 'element.normalize')(stack: test_serializer@http://localhost:8800/html/syntax/parsing/test.js:31:10
http://localhost:8800/html/syntax/parsing/test.js:162:52
step@http://localhost:8800/resources/testharness.js:798:30
onload@http://localhost:8800/html/syntax/parsing/test.js:160:13)
FAIL html5lib_tests7.html 37b910b755c2df155a3129d5a1150f0c0fdd7934 null is not an object (evaluating 'element.normalize')(stack: test_serializer@http://localhost:8800/html/syntax/parsing/test.js:31:10
http://localhost:8800/html/syntax/parsing/test.js:162:52
step@http://localhost:8800/resources/testharness.js:798:30
onload@http://localhost:8800/html/syntax/parsing/test.js:160:13)
FAIL html5lib_tests7.html 868bff3a23219b836fdc702063d637f817ce65e1 null is not an object (evaluating 'element.normalize')(stack: test_serializer@http://localhost:8800/html/syntax/parsing/test.js:31:10
http://localhost:8800/html/syntax/parsing/test.js:162:52
step@http://localhost:8800/resources/testharness.js:798:30
onload@http://localhost:8800/html/syntax/parsing/test.js:160:13)
FAIL html5lib_tests7.html a33a56f5571b4bcb23138ffb60df3824f5c53773 null is not an object (evaluating 'element.normalize')(stack: test_serializer@http://localhost:8800/html/syntax/parsing/test.js:31:10
http://localhost:8800/html/syntax/parsing/test.js:162:52
step@http://localhost:8800/resources/testharness.js:798:30
onload@http://localhost:8800/html/syntax/parsing/test.js:160:13)
FAIL html5lib_tests7.html 8ba11b54fa74a1c229d079b2902d6e33e139f33b null is not an object (evaluating 'element.normalize')(stack: test_serializer@http://localhost:8800/html/syntax/parsing/test.js:31:10
http://localhost:8800/html/syntax/parsing/test.js:162:52
step@http://localhost:8800/resources/testharness.js:798:30
onload@http://localhost:8800/html/syntax/parsing/test.js:160:13)
FAIL html5lib_tests7.html 84e2152c284f4dfee7d8d12846c08b2c025578a6 null is not an object (evaluating 'element.normalize')(stack: test_serializer@http://localhost:8800/html/syntax/parsing/test.js:31:10
http://localhost:8800/html/syntax/parsing/test.js:162:52
step@http://localhost:8800/resources/testharness.js:798:30
onload@http://localhost:8800/html/syntax/parsing/test.js:160:13)
FAIL html5lib_tests7.html 8e3432411baa59cbef731ab3ba2703cb5d518453 null is not an object (evaluating 'element.normalize')(stack: test_serializer@http://localhost:8800/html/syntax/parsing/test.js:31:10
http://localhost:8800/html/syntax/parsing/test.js:162:52
step@http://localhost:8800/resources/testharness.js:798:30
onload@http://localhost:8800/html/syntax/parsing/test.js:160:13)
FAIL html5lib_tests7.html e2f6144290512430ad25bbf9598eae77288c7b7a null is not an object (evaluating 'element.normalize')(stack: test_serializer@http://localhost:8800/html/syntax/parsing/test.js:31:10
http://localhost:8800/html/syntax/parsing/test.js:162:52
step@http://localhost:8800/resources/testharness.js:798:30
onload@http://localhost:8800/html/syntax/parsing/test.js:160:13)
FAIL html5lib_tests7.html 350ebd648764d585f4aa0c29b925e6276579e9d0 null is not an object (evaluating 'element.normalize')(stack: test_serializer@http://localhost:8800/html/syntax/parsing/test.js:31:10
http://localhost:8800/html/syntax/parsing/test.js:162:52
step@http://localhost:8800/resources/testharness.js:798:30
onload@http://localhost:8800/html/syntax/parsing/test.js:160:13)
FAIL html5lib_tests7.html 9120ef80d3ee017007f3510121ddf7eba31b79e0 null is not an object (evaluating 'element.normalize')(stack: test_serializer@http://localhost:8800/html/syntax/parsing/test.js:31:10
http://localhost:8800/html/syntax/parsing/test.js:162:52
step@http://localhost:8800/resources/testharness.js:798:30
onload@http://localhost:8800/html/syntax/parsing/test.js:160:13)
FAIL html5lib_tests7.html 2026cd3ed42e41c168dd37c8c2675584f4eef335 null is not an object (evaluating 'element.normalize')(stack: test_serializer@http://localhost:8800/html/syntax/parsing/test.js:31:10
http://localhost:8800/html/syntax/parsing/test.js:162:52
step@http://localhost:8800/resources/testharness.js:798:30
onload@http://localhost:8800/html/syntax/parsing/test.js:160:13)
FAIL html5lib_tests7.html ff2e324237e22efc8430ad7137d50d6d3d311820 null is not an object (evaluating 'element.normalize')(stack: test_serializer@http://localhost:8800/html/syntax/parsing/test.js:31:10
http://localhost:8800/html/syntax/parsing/test.js:162:52
step@http://localhost:8800/resources/testharness.js:798:30
onload@http://localhost:8800/html/syntax/parsing/test.js:160:13)
FAIL html5lib_tests7.html 02c9eb822611b0c206b544e0f2e5044695195ba8 null is not an object (evaluating 'element.normalize')(stack: test_serializer@http://localhost:8800/html/syntax/parsing/test.js:31:10
http://localhost:8800/html/syntax/parsing/test.js:162:52
step@http://localhost:8800/resources/testharness.js:798:30
onload@http://localhost:8800/html/syntax/parsing/test.js:160:13)
FAIL html5lib_tests7.html cb3d1a50dd56a85135a0856cfa1c23a091ef2af4 null is not an object (evaluating 'element.normalize')(stack: test_serializer@http://localhost:8800/html/syntax/parsing/test.js:31:10
http://localhost:8800/html/syntax/parsing/test.js:162:52
step@http://localhost:8800/resources/testharness.js:798:30
onload@http://localhost:8800/html/syntax/parsing/test.js:160:13)
FAIL html5lib_tests7.html 13847685cfff75642823a0e78c6ef232ecb9d94b null is not an object (evaluating 'element.normalize')(stack: test_serializer@http://localhost:8800/html/syntax/parsing/test.js:31:10
http://localhost:8800/html/syntax/parsing/test.js:162:52
step@http://localhost:8800/resources/testharness.js:798:30
onload@http://localhost:8800/html/syntax/parsing/test.js:160:13)
FAIL html5lib_tests7.html 99bb5e9a6e0daf62ba418dd97b5e8e3925f4137e null is not an object (evaluating 'element.normalize')(stack: test_serializer@http://localhost:8800/html/syntax/parsing/test.js:31:10
http://localhost:8800/html/syntax/parsing/test.js:162:52
step@http://localhost:8800/resources/testharness.js:798:30
onload@http://localhost:8800/html/syntax/parsing/test.js:160:13)
FAIL html5lib_tests7.html 7a8e5ec2c95e725717c564dd49bfa86c2e1a88ba null is not an object (evaluating 'element.normalize')(stack: test_serializer@http://localhost:8800/html/syntax/parsing/test.js:31:10
http://localhost:8800/html/syntax/parsing/test.js:162:52
step@http://localhost:8800/resources/testharness.js:798:30
onload@http://localhost:8800/html/syntax/parsing/test.js:160:13)
FAIL html5lib_tests7.html 17dcea170bb74d18ed4776dbb98f0bac6a11364d null is not an object (evaluating 'element.normalize')(stack: test_serializer@http://localhost:8800/html/syntax/parsing/test.js:31:10
http://localhost:8800/html/syntax/parsing/test.js:162:52
step@http://localhost:8800/resources/testharness.js:798:30
onload@http://localhost:8800/html/syntax/parsing/test.js:160:13)
FAIL html5lib_tests7.html 9457c10c9f987bbc95937b34763fe956d61d237b null is not an object (evaluating 'element.normalize')(stack: test_serializer@http://localhost:8800/html/syntax/parsing/test.js:31:10
http://localhost:8800/html/syntax/parsing/test.js:162:52
step@http://localhost:8800/resources/testharness.js:798:30
onload@http://localhost:8800/html/syntax/parsing/test.js:160:13)
FAIL html5lib_tests7.html 0fa23bb5d8b2a591afb1842b8f4c00c490c127b4 null is not an object (evaluating 'element.normalize')(stack: test_serializer@http://localhost:8800/html/syntax/parsing/test.js:31:10
http://localhost:8800/html/syntax/parsing/test.js:162:52
step@http://localhost:8800/resources/testharness.js:798:30
onload@http://localhost:8800/html/syntax/parsing/test.js:160:13)
FAIL html5lib_tests7.html f6d60b3ae48e2b69b4c25125f9b5a3ab4867521b null is not an object (evaluating 'element.normalize')(stack: test_serializer@http://localhost:8800/html/syntax/parsing/test.js:31:10
http://localhost:8800/html/syntax/parsing/test.js:162:52
step@http://localhost:8800/resources/testharness.js:798:30
onload@http://localhost:8800/html/syntax/parsing/test.js:160:13)
FAIL html5lib_tests7.html 5b0b3edcc3ce9fdc9f58eb62d326865ca0aab8c8 null is not an object (evaluating 'element.normalize')(stack: test_serializer@http://localhost:8800/html/syntax/parsing/test.js:31:10
http://localhost:8800/html/syntax/parsing/test.js:162:52
step@http://localhost:8800/resources/testharness.js:798:30
onload@http://localhost:8800/html/syntax/parsing/test.js:160:13)
